🐞 fix: component NextShape fix and export color function
ais  changed component naming and color component export to avoid repetition

Issue:  #9
